Вынес Result в отдельный файл.
